la7 & lab8 : 添加对内存管理系统的支持

+++ labcodes_answer/lab7/kern/mm/pmm.h
@@ -79,7 +79,7 @@ void print_pgdir(void);
 
 extern struct Page *pages;
 extern size_t npage;
-extern uint32_t va_pa_offset;
+extern uint64_t va_pa_offset;
 
 static inline ppn_t
 page2ppn(struct Page *page) {
